statistics/statistics-rest-api/requirements.txt
atlassian-python-api
python-dateutil
requests
plotly
jinja2
